Simplifies code base by removing unnecessary code and delegating functionality that isn't reader-specific to the factory
